feat(steam): changing behavior of steam script
Making steam script publishes the files to ~/Zomboid/Workshop
