+++ scripts/utils/index.js
@@ -2,6 +2,7 @@ const { copyFolder, moveFolder, srcPath, distPath } = require("./folders");
 const { getInfo } = require("./info");
 const { createProgressBar, startProgressBar, stopProgressBar } = require("./progressBar");
 const { patchPipeWrenchLua } = require("./patches");
+const { markdownToBbcode, extractFrontMatterData } = require("./markdown");
 
 module.exports = {
 	copyFolder,
@@ -12,5 +13,7 @@ module.exports = {
 	startProgressBar,
 	stopProgressBar,
 	getInfo,
-	patchPipeWrenchLua
+	patchPipeWrenchLua,
+	markdownToBbcode,
+	extractFrontMatterData
 };
